combined some parematers to simplify code
