--- policyengine_canada/parameters/gov/provinces/qc/tax/income/credits/cce/nondisabled_old_child_expense_limit.yaml
@@ -1,9 +1,0 @@
-description:  Quebec childcare expense limit for non-disabled old child (8-16 years old).
-values:
-  2022-01-01: 5_375
-metadata:
-  unit: currency-CAD
-  label: Quebec childcare expense limit for non-disabled old child
-  reference:
-    - title: Revenu Quebec - Calculating the Tax Credit for Childcare Expenses
-      href: https://www.revenuquebec.ca/en/citizens/tax-credits/tax-credit-for-childcare-expenses/calculating-the-tax-credit-for-childcare-expenses/
