--- policyengine_canada/parameters/gov/provinces/qc/tax/income/credits/cce/nondisabled_young_child_expense_limit.yaml
@@ -1,9 +1,0 @@
-description: Quebec caps childcare expenses at this amount for its childcare expense, for young children without disabilities.
-values:
-  2022-01-01: 10_675
-metadata:
-  unit: currency-CAD
-  label: Quebec childcare expense limit for non-disabled young child
-  reference:
-    - title: Revenu Quebec - Calculating the Tax Credit for Childcare Expenses
-      href: https://www.revenuquebec.ca/en/citizens/tax-credits/tax-credit-for-childcare-expenses/calculating-the-tax-credit-for-childcare-expenses/
