--- policyengine_canada/parameters/gov/provinces/qc/tax/income/credits/cce/young_child_age_eligibility.yaml
@@ -1,9 +1,0 @@
-description: Quebec increases the credit limit for childcare expenses for child below this age.
-values:
-  2022-01-01: 7
-metadata:
-  unit: year
-  label: Quebec children care expenses young child age eligibility
-  reference:
-  - title: Revenu Quebec - Requirements for the Tax Credit for Children Expenses
-    href: https://www.revenuquebec.ca/en/citizens/tax-credits/tax-credit-for-childcare-expenses/requirements/
