--- policyengine_canada/tests/gov/provinces/qc/tax/income/credits/cce/cce_eligible_disabled_child.yaml
@@ -1,55 +1,0 @@
-- name: The child had an infirmity and was his or her parents' own child 
-  period: 2022
-  input:
-    individual_net_income: 20_000 #no matter about income
-    is_child_of_filer: true
-    is_dependant: true
-    is_disabled: true
-    province_code: QC
-  output:
-    qc_cce_eligible_disabled_child: true
-
-- name: This child had an infirmity and was a dependant, but not a own child with income above limit
-  period: 2022
-  input:
-    individual_net_income: 20_000 #above limit
-    is_child_of_filer: false
-    is_dependant: true
-    is_disabled: true
-    province_code: QC
-  output:
-    qc_cce_eligible_disabled_child: false
-
-- name: This child had an infirmity and was a dependant, but not a own child with income below limit
-  period: 2022
-  input:
-    individual_net_income: 11_000 
-    is_child_of_filer: false
-    is_dependant: true
-    is_disabled: true
-    province_code: QC
-  output:
-    qc_cce_eligible_disabled_child: true
-
-- name: The child didn't had an infirmity but was a own child
-  period: 2022
-  input:
-    individual_net_income: 20_000 
-    is_child_of_filer: true
-    is_dependant: true
-    is_disabled: false
-    province_code: QC
-  output:
-    qc_cce_eligible_disabled_child: false
-
-
-- name: The child didn't had an infirmity but was a dependant
-  period: 2022
-  input:
-    individual_net_income: 20_000 
-    is_child_of_filer: false
-    is_dependant: true
-    is_disabled: false
-    province_code: QC
-  output:
-    qc_cce_eligible_disabled_child: false
