--- policyengine_canada/tests/gov/provinces/qc/tax/income/credits/cce/cce_eligible_nondisabled_old_child.yaml
@@ -1,15 +1,0 @@
-- name: Under age eligibility
-  period: 2022
-  input:
-    age: 16
-    province_code: QC
-  output:
-    qc_cce_eligible_nondisabled_old_child: true
-
-- name: Over age eligibility
-  period: 2022
-  input:
-    age: 18
-    province_code: QC
-  output:
-    qc_cce_eligible_nondisabled_old_child: false
