--- policyengine_canada/variables/gov/provinces/qc/tax/income/credits/cce/qc_cce_eligible_disabled_children.py
@@ -1,10 +1,0 @@
-from policyengine_canada.model_api import *
-
-
-class qc_cce_eligible_disabled_children(Variable):
-    value_type = int
-    entity = Household
-    label = "Quebec childcare tax credit eligible disabled children"
-    definition_period = YEAR
-
-    adds = ["qc_cce_eligible_disabled_child"]
